+++ bench/__init__.py
@@ -1,0 +1,2 @@
+# Initialize as subpackage
+__all__ = []
